pyproject.toml
[project]
name = "nba-intel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.85",
    "chromadb>=1.5.6",
    "dbt-core>=1.11.7",
    "dbt-duckdb>=1.10.1",
    "duckdb>=1.5.1",
    "fastapi>=0.135.3",
    "groq>=1.1.2",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.1.0",
    "mlflow>=3.11.1",
    "nba-api>=1.11.4",
    "pandas>=2.3.3",
    "pyarrow>=23.0.1",
    "streamlit>=1.56.0",
    "uvicorn>=0.44.0",
]
